test(PersonNameTest): 사람 이름에 대한 검증 테스트 코드 위치 변경
