+++ src/test/java/model/PersonTest.java
@@ -1,12 +1,9 @@
 package model;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 class PersonTest {
 
@@ -16,13 +13,5 @@ class PersonTest {
         String name = "moly";
         Person person = new Person(name);
         assertThat(person.getName()).isEqualTo(name);
-    }
-
-    @ParameterizedTest(name = "이름은 최소 1글자 최대 5글자다.")
-    @ValueSource(strings = {"", "mollly"})
-    void createPersonThrowException(String name) {
-        assertThatThrownBy(() -> new Person(name))
-                .isExactlyInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이름은 최소 1글자 최대 5글자여야 합니다.");
     }
 }
